+++ lib/metrics/pc_error_wapper.py
@@ -12,7 +12,7 @@ _DIVIDERS = ['1. Use infile1 (A) as reference, loop over A, use normals on B. (A
 
 
 def mpeg_pc_error(
-        infile1: str, infile2: str, resolution: int, normal_file: str = '',
+        infile1: str, infile2: str, resolution: float, normal_file: str = '',
         hausdorff: bool = False, color: bool = False, threads: int = 1, command='',
         hooks: Tuple[Callable[[str], Tuple[str, Union[None, int, float, str]]]] = ()
 ) -> Dict[str, float]:
